requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-environ==0.10.0
djangorestframework==3.14.0
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
